Changed to a pixel buffer method

+++ slib.h
@@ -134,6 +134,6 @@ namespace slib
     
     struct Color
     {
-        float r, g, b;
+        int r, g, b;
     };
 }
